requirements.txt
anthropic
rich
python-dotenv